requirements.txt
google-cloud-storage==1.43.0
beautifulsoup4==4.9.3
flanker==0.9.11